+++ nonUnity_folder/rayCoordinates.R
@@ -1,8 +1,8 @@
 # WD
-setwd("~/Unity/UXF_tracker_sampleProject/input_data")
+setwd("E:/research_projects/OLM_project/experiment/unityProjects/DiamondWorld/nonUnity_folder")
 
 # Params
-totalNumberOfRays <- 100
+totalNumberOfRays <- 25
 margin            <- 0.05
 
 raysPerDim        <- sqrt(totalNumberOfRays)
@@ -16,7 +16,6 @@ y <- x
 # Combine all possibilities
 rayCoordinates <- expand.grid(x, y)
 
-write.table(rayCoordinates, 'rayCoordinates.txt', quote = FALSE, sep = '\t', row.names = FALSE, col.names = FALSE)
 
 
 # Lib
@@ -24,3 +23,6 @@ library(ggplot2)
 
 
 ggplot(rayCoordinates, aes(x = Var1, y = Var2)) + geom_point()
+
+paste0("[", paste0(rayCoordinates$Var1, collapse = ", "), "]")
+paste0("[", paste0(rayCoordinates$Var2, collapse = ", "), "]")