+++ change_log.md
@@ -73,4 +73,5 @@
     - Baked lights even if benefit is not entirely clear. 
     - Bakad again but didn't remove the reflection probes as it seems important for the objects rendering correctly. 
     - Clean up stuff from old version like prefabs and samples from UXF.
-    - Number of small changes: 1. Renamed destroyAllCubesAtTrialEnd function, 2. Update file logo
+    - Number of small changes: 1. Renamed destroyAllCubesAtTrialEnd function, 2. Update file logo, 3. add object index to log. 
+    - Create new basic layout with minimum distance betwee objects.